# mall_agent_openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mall Assistant Data API",
    "description": "Provides real-time information about mall deals, store offers, events, user history, directions, and the full store/dining directory.",
    "version": "1.4.0"
  },
  "servers": [
    {
      "url": "https://mall-agent-production.up.railway.app/"
    }
  ],
  "paths": {
    "/deals": {
      "get": {
        "summary": "Get Mall Deals",
        "operationId": "getDeals",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/events": {
      "get": {
        "summary": "Get Mall Events",
        "operationId": "getEvents",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/stores": {
      "get": {
        "summary": "Get Store Directory",
        "description": "Search for fashion, sports, and accessories stores by floor or category.",
        "operationId": "getStores",
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "floor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/dining": {
      "get": {
        "summary": "Get Dining Directory",
        "description": "Search for food court, premium dining, and quick service outlets by floor or cuisine.",
        "operationId": "getDining",
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "cuisine", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/user": {
      "get": {
        "summary": "Find User",
        "operationId": "getUser",
        "parameters": [
          { "name": "identifier", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Register User",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" } } }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/orders": {
      "get": {
        "summary": "Get Orders",
        "operationId": "getOrders",
        "parameters": [
          { "name": "phone_number", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/directions": {
      "get": {
        "summary": "Get Mall Directions (Driving & Walking)",
        "operationId": "getDirections",
        "parameters": [
          { "name": "origin", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "destination", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    }
  },
  "components": {
    "schemas": {
      "UserCreate": {
        "type": "object",
        "required": ["name", "phone_number", "email"],
        "properties": {
          "name": { "type": "string" },
          "phone_number": { "type": "string" },
          "email": { "type": "string" },
          "last_activity": { "type": "string" }
        }
      }
    }
  }
}